refactor: extract remaining search helpers
