+++ src/core/src/service/search/datafusion/mod.rs
@@ -13,8 +13,6 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-use std::str::FromStr;
-
 pub mod distributed_plan;
 pub mod exec;
 pub mod merge;
@@ -23,61 +21,6 @@ pub mod plan;
 pub use search::datafusion::table_provider;
 pub mod udf;
 
-pub use search::datafusion::{peak_memory_pool, plan_metrics, planner, storage, udaf};
-
-#[derive(PartialEq, Debug)]
-pub enum MemoryPoolType {
-    Greedy,
-    Fair,
-    None,
-}
-
-impl FromStr for MemoryPoolType {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().as_str() {
-            "greedy" | "" => Ok(MemoryPoolType::Greedy),
-            "fair" => Ok(MemoryPoolType::Fair),
-            "none" | "off" => Ok(MemoryPoolType::None),
-            _ => Err(format!("Invalid memory pool type '{s}'")),
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_memory_pool_type_from_str_greedy() {
-        assert_eq!(
-            "greedy".parse::<MemoryPoolType>(),
-            Ok(MemoryPoolType::Greedy)
-        );
-        assert_eq!(
-            "GREEDY".parse::<MemoryPoolType>(),
-            Ok(MemoryPoolType::Greedy)
-        );
-        assert_eq!("".parse::<MemoryPoolType>(), Ok(MemoryPoolType::Greedy));
-    }
-
-    #[test]
-    fn test_memory_pool_type_from_str_fair() {
-        assert_eq!("fair".parse::<MemoryPoolType>(), Ok(MemoryPoolType::Fair));
-        assert_eq!("FAIR".parse::<MemoryPoolType>(), Ok(MemoryPoolType::Fair));
-    }
-
-    #[test]
-    fn test_memory_pool_type_from_str_none() {
-        assert_eq!("none".parse::<MemoryPoolType>(), Ok(MemoryPoolType::None));
-        assert_eq!("off".parse::<MemoryPoolType>(), Ok(MemoryPoolType::None));
-        assert_eq!("OFF".parse::<MemoryPoolType>(), Ok(MemoryPoolType::None));
-    }
-
-    #[test]
-    fn test_memory_pool_type_from_str_invalid() {
-        assert!("unknown".parse::<MemoryPoolType>().is_err());
-        assert!("pool".parse::<MemoryPoolType>().is_err());
-    }
-}
+pub use search::datafusion::{
+    MemoryPoolType, peak_memory_pool, plan_metrics, planner, storage, udaf,
+};
