+++ src/search/src/cache/result_utils.rs
