+++ src/search/src/datafusion/mod.rs
@@ -23,3 +23,61 @@ pub mod storage;
 pub mod table_provider;
 pub mod udaf;
 pub mod udf;
+
+#[derive(PartialEq, Debug)]
+pub enum MemoryPoolType {
+    Greedy,
+    Fair,
+    None,
+}
+
+impl FromStr for MemoryPoolType {
+    type Err = String;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s.to_lowercase().as_str() {
+            "greedy" | "" => Ok(MemoryPoolType::Greedy),
+            "fair" => Ok(MemoryPoolType::Fair),
+            "none" | "off" => Ok(MemoryPoolType::None),
+            _ => Err(format!("Invalid memory pool type '{s}'")),
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_memory_pool_type_from_str_greedy() {
+        assert_eq!(
+            "greedy".parse::<MemoryPoolType>(),
+            Ok(MemoryPoolType::Greedy)
+        );
+        assert_eq!(
+            "GREEDY".parse::<MemoryPoolType>(),
+            Ok(MemoryPoolType::Greedy)
+        );
+        assert_eq!("".parse::<MemoryPoolType>(), Ok(MemoryPoolType::Greedy));
+    }
+
+    #[test]
+    fn test_memory_pool_type_from_str_fair() {
+        assert_eq!("fair".parse::<MemoryPoolType>(), Ok(MemoryPoolType::Fair));
+        assert_eq!("FAIR".parse::<MemoryPoolType>(), Ok(MemoryPoolType::Fair));
+    }
+
+    #[test]
+    fn test_memory_pool_type_from_str_none() {
+        assert_eq!("none".parse::<MemoryPoolType>(), Ok(MemoryPoolType::None));
+        assert_eq!("off".parse::<MemoryPoolType>(), Ok(MemoryPoolType::None));
+        assert_eq!("OFF".parse::<MemoryPoolType>(), Ok(MemoryPoolType::None));
+    }
+
+    #[test]
+    fn test_memory_pool_type_from_str_invalid() {
+        assert!("unknown".parse::<MemoryPoolType>().is_err());
+        assert!("pool".parse::<MemoryPoolType>().is_err());
+    }
+}
+use std::str::FromStr;
