+++ src/search/src/lib.rs
@@ -14,6 +14,7 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 pub mod bloom_pruner;
+pub mod cache;
 pub mod datafusion;
 pub mod enrichment;
 pub mod index;
